+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,9 @@
 
 ## Next Release
 
+- \[engine\] Add `transaction` to `WITHDRAWAL_RESOLVED` event
+- \[node\] Make `transaction` a non-optional return value
+
 ## 0.2.0-beta.8
 
 - \[engine, router\] Allow users to specify if they want to receive an exact amount when requesting a quote